+++ application/controllers/Example.php
@@ -7,22 +7,34 @@ class Example extends CI_Controller {
 	{
 		parent::__construct();
 
+		// Load library and url helper
 		$this->load->library('facebook');
 		$this->load->helper('url');
 	}
 
+	// ------------------------------------------------------------------------
+
+	/**
+	 * Index page
+	 */
 	public function index()
 	{
 		$this->load->view('examples/start');
 	}
 
+	// ------------------------------------------------------------------------
+
+	/**
+	 * Web redirect login example page
+	 */
 	public function web_login()
 	{
-
 		$data['user'] = array();
 
+		// Check if user is logged in
 		if ($this->facebook->is_authenticated())
 		{
+			// User logged in, get user details
 			$user = $this->facebook->request('get', '/me');
 			if (!isset($user['error']))
 			{
@@ -30,24 +42,41 @@ class Example extends CI_Controller {
 			}
 		}
 
+		// display view
 		$this->load->view('examples/web', $data);
 	}
 
+	// ------------------------------------------------------------------------
+
+	/**
+	 * JS SDK login example
+	 */
 	public function js_login()
 	{
+		// Load view
 		$this->load->view('examples/js');
 	}
 
+	// ------------------------------------------------------------------------
+
+	/**
+	 * AJAX request method for positing to facebook feed
+	 */
 	public function post()
 	{
-
 		header('Content-Type: application/json');
 
 		$result = $this->facebook->request('post', '/me/feed', array('message' => $this->input->post('message')));
 		echo json_encode($result);
-
 	}
 
+	// ------------------------------------------------------------------------
+
+	/**
+	 * Logout for web redirect example
+	 *
+	 * @return  [type]  [description]
+	 */
 	public function logout()
 	{
 		$this->facebook->destroy_session();