+++ changelog.md
@@ -1,4 +1,9 @@
-# Changelog for Facebook PHP SDK v4 for CodeIgniter
+# Changelog for Facebook PHP SDK for CodeIgniter
+
+#### 3.0.0-b01 (October 2015)
+- NEW: Support for Facebook PHP SDK v5.
+- NEW: Now using one singly dynamic method to do any type of Facebook Graph request.
+- CHANGE: Dropped support for Facebook PHP SDK v4.
 
 #### 2.0.0 (April 2015)
 - NEW: Added support for Redirect login.